requirements.txt
streamlit
tensorflow
scikit-learn
joblib
numpy
Pillow
xgboost
catboost
lightgbm
streamlit-webrtc
